summariser/requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
transformers
torch==1.10.0+cu113 
sentencepiece
grpcio
grpcio_tools